make get_output_workspace usable by dash
