+++ geoapps/clustering/application.py
@@ -606,9 +606,10 @@ class Clustering(ScatterPlots):
             )
 
             temp_geoh5 = f"Clustering_{time():.0f}.geoh5"
-            with self.get_output_workspace(
-                self.export_directory.selected_path, temp_geoh5
-            ) as workspace:
+            ws, self.live_link.value = self.get_output_workspace(
+                self.live_link.value, self.export_directory.selected_path, temp_geoh5
+            )
+            with ws as workspace:
                 obj = obj.copy(parent=workspace)
                 cluster_groups = obj.add_data(
                     {
