+++ geoapps/contours/application.py
@@ -160,9 +160,10 @@ class ContourValues(PlotSelection2D):
     def trigger_click(self, _):
         param_dict = self.get_param_dict()
         temp_geoh5 = f"{string_name(param_dict['export_as'])}_{time():.0f}.geoh5"
-        with self.get_output_workspace(
-            self.export_directory.selected_path, temp_geoh5
-        ) as workspace:
+        ws, self.live_link.value = self.get_output_workspace(
+            self.live_link.value, self.export_directory.selected_path, temp_geoh5
+        )
+        with ws as workspace:
             for key, value in param_dict.items():
                 if isinstance(value, ObjectBase):
                     param_dict[key] = value.copy(parent=workspace, copy_children=True)
