+++ geoapps/octree_creation/application.py
@@ -227,9 +227,10 @@ class OctreeMesh(ObjectDataSelection):
                 continue
 
         temp_geoh5 = f"{self.ga_group_name.value}_{time():.0f}.geoh5"
-        with self.get_output_workspace(
-            self.export_directory.selected_path, temp_geoh5
-        ) as new_workspace:
+        ws, self.live_link.value = self.get_output_workspace(
+            self.live_link.value, self.export_directory.selected_path, temp_geoh5
+        )
+        with ws as new_workspace:
 
             for key, value in param_dict.items():
                 if isinstance(value, ObjectBase):
