+++ geoapps/peak_finder/application.py
@@ -1299,9 +1299,11 @@ class PeakFinder(ObjectDataSelection):
                 param_dict[name] = group[member]
 
         temp_geoh5 = f"{self.ga_group_name.value}_{time():.0f}.geoh5"
-        with self.get_output_workspace(
-            self.export_directory.selected_path, temp_geoh5
-        ) as new_workspace:
+
+        ws, self.live_link.value = self.get_output_workspace(
+            self.live_link.value, self.export_directory.selected_path, temp_geoh5
+        )
+        with ws as new_workspace:
 
             for key, value in param_dict.items():
                 if isinstance(value, ObjectBase):
